Fixes after enabled compiler warnings

+++ src/Fg/Util/FileUtil.cpp
@@ -37,7 +37,8 @@ namespace util
 // Writing the buffer to a .PPM file, assuming it has RGB-structure, one byte
 // per color component.
 // http://www.antigrain.com/doc/basic_renderers/basic_renderers.agdoc.html
-bool writePpmFile(const unsigned char* buf,
+bool writePpmFile(
+    const unsigned char* buf,
     unsigned width,
     unsigned height,
     unsigned bytePerPixel,
@@ -61,11 +62,14 @@ bool writePpmFile(const unsigned char* buf,
 }
 
 //https://stackoverflow.com/a/37575457
-bool compareFiles(const fg::filesystem::path& filePath1,
+bool compareFiles(
+    const fg::filesystem::path& filePath1,
     const fg::filesystem::path& filePath2)
 {
-  std::ifstream f1(filePath1.string(), std::ifstream::binary | std::ifstream::ate);
-  std::ifstream f2(filePath2.string(), std::ifstream::binary | std::ifstream::ate);
+  std::ifstream f1(
+      filePath1.string(), std::ifstream::binary | std::ifstream::ate);
+  std::ifstream f2(
+      filePath2.string(), std::ifstream::binary | std::ifstream::ate);
 
   if(f1.fail() || f2.fail()) {
     return false;  // File problem.
@@ -78,7 +82,8 @@ bool compareFiles(const fg::filesystem::path& filePath1,
   // Seek back to beginning and use std::equal to compare contents.
   f1.seekg(0, std::ifstream::beg);
   f2.seekg(0, std::ifstream::beg);
-  return std::equal(std::istreambuf_iterator<char>(f1.rdbuf()),
+  return std::equal(
+      std::istreambuf_iterator<char>(f1.rdbuf()),
       std::istreambuf_iterator<char>(),
       std::istreambuf_iterator<char>(f2.rdbuf()));
 }
@@ -87,17 +92,18 @@ bool compareFiles(const fg::filesystem::path& filePath1,
 // https://stackoverflow.com/a/43009155
 std::string readFile(const fg::filesystem::path& fileName)
 {
-  std::ifstream ifs(fileName.string(), std::ios::in | std::ios::binary | std::ios::ate);
+  std::ifstream ifs(
+      fileName.string(), std::ios::in | std::ios::binary | std::ios::ate);
 
   std::ifstream::pos_type fileSize = ifs.tellg();
   if(fileSize < 0)
     return std::string();
 
   ifs.seekg(0, std::ios::beg);
-  std::vector<char> bytes(fileSize);
+  std::vector<char> bytes(static_cast<std::size_t>(fileSize));
   ifs.read(&bytes[0], fileSize);
 
-  return std::string(&bytes[0], fileSize);
+  return std::string(&bytes[0], static_cast<std::size_t>(fileSize));
 }
 
 }  // namespace util
